export/requirements.txt
apscheduler==3.10.4
brother-ql==0.9.4
email-validator==2.1.0
flask==3.0.0
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
pillow==10.1.0
psycopg2-binary==2.9.9
woocommerce==3.0.0